+++ src/presentation/routes/home/components/musicList/index.tsx
@@ -2,32 +2,30 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { MusicPanel } from './music';
 import { Music } from 'domain/entity/music';
-import { SebaMusicState } from 'presentation/redux/getMusicInfo/reducer';
 import { FetchResult } from 'presentation/redux/actionTypes';
 import { FetchStatus } from 'presentation/redux/FetchStatus';
 import { Loading } from 'presentation/components/loading';
+import { List } from 'immutable';
 const styles = require('./styles.scss');
 
-type StateProps = {
-  sebaMusic: SebaMusicState;
-  sebaMusicFetchState: FetchResult;
+type Props = {
+  musicList: List<Music>;
+  musicFetchState: FetchResult;
 };
 
-type Props = StateProps;
-
-const MusicList: React.FC<Props> = ({
-  sebaMusic,
-  sebaMusicFetchState
+export const MusicList: React.FC<Props> = ({
+  musicList,
+  musicFetchState
 }) => { 
   const renderMusicComponent = () => {
-    switch (sebaMusicFetchState.fetchState) {
+    switch (musicFetchState.fetchState) {
       case FetchStatus.DEFAULT:
       case FetchStatus.LOADING:
         return <Loading />;
       case FetchStatus.ERROR:
         return <div>Error</div>;
       case FetchStatus.SUCCESS:
-        return sebaMusic.data.map((music: Music) => <MusicPanel musicInfo={music} />);
+        return musicList.map((music: Music) => <MusicPanel musicInfo={music} />);
       default:
         return;
     }
@@ -44,12 +42,3 @@ const MusicList: React.FC<Props> = ({
 
 
 }
-
-const mapStateToProps = (state: StateProps): StateProps => ({
-  sebaMusic: state.sebaMusic,
-  sebaMusicFetchState: state.sebaMusicFetchState,
-});
-
-export default connect(
-  mapStateToProps
-)(MusicList);
