+++ src/presentation/routes/home/container/index.tsx
@@ -2,13 +2,24 @@ import React, { useEffect } from 'react';
 import { RequestSebaMusic, requestSebaCurationMusicLists } from 'presentation/redux/getMusicInfo/actions';
 import { Dispatch, bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import MusicList from 'presentation/routes/home/components/musicList/index';
+import { MusicList } from 'presentation/routes/home/components/musicList/index';
+import { SebaMusicState } from 'presentation/redux/getMusicInfo/reducer';
+import { FetchResult } from 'presentation/redux/actionTypes';
 
 type DispatchProps = {
   requestSebaCurationMusicLists: RequestSebaMusic;
 };
 
-const SebaMusicContainer: React.FC<DispatchProps> = ({
+type StateProps = {
+  sebaMusic: SebaMusicState;
+  sebaMusicFetchState: FetchResult;
+};
+
+type Props = DispatchProps & StateProps;
+
+const SebaMusicContainer: React.FC<Props> = ({
+  sebaMusic,
+  sebaMusicFetchState,
   requestSebaCurationMusicLists
 }) => {
   useEffect(() => {
@@ -17,7 +28,10 @@ const SebaMusicContainer: React.FC<DispatchProps> = ({
     requestSebaCurationMusicLists
   ])
   return(
-    <MusicList />
+    <MusicList 
+      musicList={sebaMusic.data}
+      musicFetchState={sebaMusicFetchState}
+    />
   )
 }
 const mapDispatchToProps = (dispatch: Dispatch) =>
@@ -26,8 +40,12 @@ const mapDispatchToProps = (dispatch: Dispatch) =>
     },
     dispatch,
   );
+const mapStateToProps = (state: StateProps): StateProps => ({
+  sebaMusic: state.sebaMusic,
+  sebaMusicFetchState: state.sebaMusicFetchState,
+});
 
 export default connect(
-  null,
+  mapStateToProps,
   mapDispatchToProps,
 )(SebaMusicContainer);
